requirements.txt
APScheduler==3.10.4
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.1.3
